+++ packages/app/src/components/time-series-chart/components/tooltip/tooltip-series-list.tsx
@@ -3,17 +3,17 @@ import {
   isDateValue,
   TimestampedValue,
 } from '@corona-dashboard/common';
-import { useIntl } from '~/intl';
 import css from '@styled-system/css';
 import { ReactNode } from 'react';
 import styled from 'styled-components';
 import { Box } from '~/components/base';
 import { InlineText, Text } from '~/components/typography';
 import { VisuallyHidden } from '~/components/visually-hidden';
+import { useIntl } from '~/intl';
+import { colors } from '~/style/theme';
 import { SeriesConfig, useFormatSeriesValue } from '../../logic';
 import { SeriesIcon } from '../series-icon';
 import { TooltipData } from './types';
-import { colors } from '~/style/theme';
 
 export function TooltipSeriesList<T extends TimestampedValue>({
   data: tooltipData,
@@ -61,7 +61,7 @@ export function TooltipSeriesList<T extends TimestampedValue>({
     <section>
       <VisuallyHidden>{dateString}</VisuallyHidden>
 
-      {timespanAnnotation && (
+      {timespanAnnotation && timespanAnnotation.type !== 'divider' && (
         <Text fontSize={0} color={colors.annotation} textAlign={'center'}>
           {timespanAnnotation.shortLabel || timespanAnnotation.label}
         </Text>
